task-132: fix adding product with selected category to private list
